refactor(api): extract user service interface for extension
